% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_pack.R
\name{cq}
\alias{cq}
\title{Unquoted String Concatenation}
\usage{
cq(...)
}
\arguments{
\item{...}{Unqouted strings}
}
\value{
A character string vector
}
\description{
A convenience function.  Allows a character string
vector to be generated without typing the quotation
marks.  Very similar to Hmisc::Cs
}
\examples{
packages <- cq(tidyverse, readxl, lme4, mgcv, splines, randomForest, rpart)
packages
}
