% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requireData.R
\name{requireData}
\alias{requireData}
\title{Attach Package and Data}
\usage{
requireData(
  package = stop("you must specify a package"),
  lib.loc = NULL,
  quietly = TRUE,
  character.only = FALSE,
  warn.conflicts = TRUE,
  reallyQuietly = quietly,
  ...
)
}
\arguments{
\item{package}{The name of the package}

\item{lib.loc}{as for base::require()}

\item{quietly}{as for base::require()}

\item{character.only}{as for base::require()}

\item{warn.conflicts}{as for base::require()}

\item{reallyQuietly}{logical: suppress startup messages?}

\item{...}{extra arguments, currently ignored.}
}
\value{
logical: as for base::require()
}
\description{
Attaches a package and its dependencies to the search path
including any data sets if the package itself, or any of its
dependencies, have data sets available, but lazyLoad set to false
}
\examples{
requireData(lazyData)
}
