% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_pack.R
\name{use_packages}
\alias{use_packages}
\title{Add packages to the search path}
\usage{
use_packages(..., list = NULL, quietly = TRUE, show_data = TRUE)
}
\arguments{
\item{...}{names of packages required as unquoted character strings}

\item{list}{package names as a character string vector (or list)}

\item{quietly}{logical: should the package startup messages be stifled?}

\item{show_data}{logical: should non-lazy-loaded data sets be exposed?}
}
\value{
A character string vector of the package names, invisibly
}
\description{
Make multiple packages available on the search path,
optionally installing any which are not yet installed,
(provided they are available on the default repository).
}
\examples{
\dontrun{
modelling <- cq(lme4, splines, mgcv, randomForest, rpart)
use_packages(tidyverse, readxl, list = modelling, quietly = TRUE)
}
}
